CocoEvaluator: Fix bug on tokenizationg

+++ language_evaluation/coco_caption_py3/pycocoevalcap/tokenizer/ptbtokenizer.py
@@ -23,38 +23,43 @@ PUNCTUATIONS = ["''", "'", "``", "`", "-LRB-", "-RRB-", "-LCB-", "-RCB-", \
 
 class PTBTokenizer:
     """Python wrapper of Stanford PTBTokenizer"""
+    def __init__(self, tokenization_fn=None):
+        self.tokenization_fn = tokenization_fn
 
     def tokenize(self, captions_for_image):
         cmd = ['java', '-cp', STANFORD_CORENLP_3_4_1_JAR, \
                 'edu.stanford.nlp.process.PTBTokenizer', \
                 '-preserveLines', '-lowerCase']
 
-        # ======================================================
-        # prepare data for PTB Tokenizer
-        # ======================================================
-        final_tokenized_captions_for_image = {}
-        image_id = [k for k, v in captions_for_image.items() for _ in range(len(v))]
-        sentences = '\n'.join([c['caption'].replace('\n', ' ') for k, v in captions_for_image.items() for c in v])
+        if self.tokenization_fn:
+            raise ValueError("Custom tokenizer is not supported yet")
+        else:
+            # ======================================================
+            # prepare data for PTB Tokenizer
+            # ======================================================
+            final_tokenized_captions_for_image = {}
+            image_id = [k for k, v in captions_for_image.items() for _ in range(len(v))]
+            sentences = '\n'.join([c['caption'].replace('\n', ' ') for k, v in captions_for_image.items() for c in v])
 
-        # ======================================================
-        # save sentences to temporary file
-        # ======================================================
-        path_to_jar_dirname=os.path.dirname(os.path.abspath(__file__))
-        tmp_file = tempfile.NamedTemporaryFile(delete=False, dir=path_to_jar_dirname)
-        tmp_file.write(sentences.encode())
-        tmp_file.close()
+            # ======================================================
+            # save sentences to temporary file
+            # ======================================================
+            path_to_jar_dirname=os.path.dirname(os.path.abspath(__file__))
+            tmp_file = tempfile.NamedTemporaryFile(delete=False, dir=path_to_jar_dirname)
+            tmp_file.write(sentences.encode())
+            tmp_file.close()
 
-        # ======================================================
-        # tokenize sentence
-        # ======================================================
-        cmd.append(os.path.basename(tmp_file.name))
-        p_tokenizer = subprocess.Popen(cmd, cwd=path_to_jar_dirname, \
-                stdout=subprocess.PIPE)
-        token_lines = p_tokenizer.communicate(input=sentences.rstrip())[0]
-        token_lines = token_lines.decode()
-        lines = token_lines.split('\n')
-        # remove temp file
-        os.remove(tmp_file.name)
+            # ======================================================
+            # tokenize sentence
+            # ======================================================
+            cmd.append(os.path.basename(tmp_file.name))
+            p_tokenizer = subprocess.Popen(cmd, cwd=path_to_jar_dirname, \
+                    stdout=subprocess.PIPE)
+            token_lines = p_tokenizer.communicate(input=sentences.rstrip())[0]
+            token_lines = token_lines.decode()
+            lines = token_lines.split('\n')
+            # remove temp file
+            os.remove(tmp_file.name)
 
         # ======================================================
         # create dictionary for tokenized captions
